Refactor index.js
created a router for Task endpoints
created a router for User endpoints
glued them in the index.js
