+++ src/index.js
@@ -5,163 +5,18 @@ require('./db/mongoose')
 
 const User = require('./model/user')
 const Task = require('./model/task')
+const userRouter = require('./routers/user')
+const taskRouter = require('./routers/task')
 
 const app = express()
 const port = process.env.PORT || 3000
 
 app.use(express.json())
 
-// fetch multiple users route handler
-app.get('/users', async (req, res) => {
-    try {
-        const users = await User.find({})
-        res.send(users)
-    } catch (e) {
-        res.status(500).send(e)
-    }
-})
-
-app.get('/users/:id', async (req, res) => {
-
-    const _id = req.params.id
-
-    if (!mongoose.isValidObjectId(_id)){
-        return res.status(500).send('Invalide object ID')
-    }
-
-    try {
-        const user = await User.findById(_id)
-        if (!user) {
-            return res.status(404).send()
-        }
-        res.send(user)
-    } catch (e) {
-        res.status(500).send(e)
-    }
-})
-
-app.post('/users' , async (req , res)=>{
-    const user = new User(req.body)
-    try {
-        await user.save()
-        res.status(201).send(user)
-    } catch (e) {
-        res.status(400).send(e)
-    }
-})
-
-app.patch('/users/:id', async (req, res) => {
-    
-    const updates = Object.keys(req.body)
-    const allowedUpdates = ['name', 'email', 'password', 'age']
-    const isValidOperation = updates.every((update) => allowedUpdates.includes(update))
-
-    if (!isValidOperation) {
-        return res.status(400).send({ error: 'Invalid updates!'})
-    }
-
-    try {
-        const user = await User.findByIdAndUpdate(req.params.id, req.body, {new: true, runValidators: true})
-        if (!user) {
-            return res.status(404).send()
-        }
-        res.send(user)
-    } catch (e) {
-        res.status(400).send(e)
-    }
-})
-
-app.delete('/users/:id', async (req, res) => {
-    try {
-        const user = await User.findByIdAndDelete(req.params.id)
-        if (!user) {
-            return res.status(404).send()
-        }
-        res.send(user)
-    } catch(e) {
-        res.status(500).send(e)
-    }
-})
-
-app.post('/tasks', async (req, res) => {
-    const task = new Task(req.body)
-    try {
-        await task.save()
-        res.status(201).send(task)
-    } catch (e) {
-        res.status(400).send(e)
-    }
-})
-
-app.patch('/tasks/:id', async (req, res) => {
-    const updates = Object.keys(req.body)
-    const allowedUpdates = ['description', 'completed']
-    const isValidOperation = updates.every((update) => allowedUpdates.includes(update))
-
-    if (!isValidOperation) {
-        return res.status(400).send({ error: 'Invalid updates!'})
-    }
-
-    try {
-        const task = await Task.findByIdAndUpdate(req.params.id, req.body, { new: true, runValidators: true })
-        if (!task) {
-            return res.status(404).send()
-        }
-        res.send(task)
-    } catch (e) {
-        res.status(400).send(e)
-    }
-})
-
-// fetch multiple tasks route handler
-app.get('/tasks' , async (req , res) => {
-    try {
-        const tasks = await Task.find({})
-        res.send(tasks)
-    } catch (e) {
-        res.status(500).send(e)
-    }
-})
-
-// fetch a specific task by _id route handler
-app.get('/tasks/:id', async (req, res) => {
-
-    const _id = req.params.id
-
-    if (!mongoose.isValidObjectId(_id)) {
-        return res.status(500).send('Invalid object ID')
-    }
-
-    try {
-        const task = await Task.findById(_id)
-        if (!task) {
-            return res.status(404).send()
-        }
-        res.send(task)
-    } catch (e) {
-        res.status(500).send(e)
-    }
-
-});
+// TODO: CREATE A ROUTER
+app.use(userRouter)
+app.use(taskRouter)
 
 
-// Goal: Allow for removal of tasks
-// 
-// 1. Setup the endpoint handler
-// 2. Attempt to delete the task by id
-//      - Handle success
-//      - Handle task not found
-//      - Handle error
-// 3. Test your work
-app.delete('/tasks/:id', async (req, res) => {
-    try {
-        const task = await Task.findByIdAndDelete(req.params.id)
-        if (!task) {
-            return res.status(404).send()
-        }
-        res.send(task)
-    } catch (e) {
-        res.status(500).send(e)
-    }
-})
+
 app.listen(port, () => console.log(`Example app listening on port ${port}!`))
