+++ src/routers/user.js
@@ -1,0 +1,86 @@
+const express = require('express');
+const mongoose = require('mongoose')
+// require('../db/mongoose')
+
+const User = require('../model/user')
+const router = express.Router();
+
+router.get('/test', (req, res) => {
+    res.send('Router from a new file')
+})
+
+
+// fetch multiple users route handler
+router.get('/users', async (req, res) => {
+    try {
+        const users = await User.find({})
+        res.send(users)
+    } catch (e) {
+        res.status(500).send(e)
+    }
+})
+
+router.get('/users/:id', async (req, res) => {
+
+    const _id = req.params.id
+
+    if (!mongoose.isValidObjectId(_id)){
+        return res.status(500).send('Invalide object ID')
+    }
+
+    try {
+        const user = await User.findById(_id)
+        if (!user) {
+            return res.status(404).send()
+        }
+        res.send(user)
+    } catch (e) {
+        res.status(500).send(e)
+    }
+})
+
+router.post('/users' , async (req , res)=>{
+    const user = new User(req.body)
+    try {
+        await user.save()
+        res.status(201).send(user)
+    } catch (e) {
+        res.status(400).send(e)
+    }
+})
+
+router.patch('/users/:id', async (req, res) => {
+    
+    const updates = Object.keys(req.body)
+    const allowedUpdates = ['name', 'email', 'password', 'age']
+    const isValidOperation = updates.every((update) => allowedUpdates.includes(update))
+
+    if (!isValidOperation) {
+        return res.status(400).send({ error: 'Invalid updates!'})
+    }
+
+    try {
+        const user = await User.findByIdAndUpdate(req.params.id, req.body, {new: true, runValidators: true})
+        if (!user) {
+            return res.status(404).send()
+        }
+        res.send(user)
+    } catch (e) {
+        res.status(400).send(e)
+    }
+})
+
+router.delete('/users/:id', async (req, res) => {
+    try {
+        const user = await User.findByIdAndDelete(req.params.id)
+        if (!user) {
+            return res.status(404).send()
+        }
+        res.send(user)
+    } catch(e) {
+        res.status(500).send(e)
+    }
+})
+
+
+module.exports = router
